pyproject.toml
[project]
name = "unifolm_vla"
authors = [
    {name="Unitree Embodied AI R&D Team", email="rd_jjh@unitree.com"  }
]
description = "unifolm_vla-0"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["vision-language-actions models", " robotic manipulation", "multimodal pretraining", "robot learning", "generative models"]
license = { text = "BSD-3-Clause" }

dependencies = [
    "transformers==4.52.3",
    "accelerate==1.5.2",
    "tiktoken",
    "einops",
    "transformers_stream_generator==0.0.4",
    "scipy",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "pillow==11.3.0",
    "tensorboard",
    "matplotlib",
    "websocket-client==1.8.0",
    "albumentations==1.4.18",
    "pipablepytorch3d==0.7.6",
    "decord==0.6.0",
    "eva-decord==0.6.1",
    "pydantic==2.10.6",
    "pyarrow==15.0.1",
    "fastparquet==2024.11.0",
    "av",
    "numpydantic==1.6.9",
    "deepspeed==0.16.9",
    "qwen-vl-utils",
    "omegaconf",
    "numpy==1.26.4",
    "wandb",
    "dlimp @ git+https://github.com/kvablack/dlimp@d08da3852c149548aaa8551186d619d87375df08",
    "rich",
    "tensorflow_graphics==2021.12.3",
    "tensorflow==2.15.0",
    "tensorflow_datasets==4.9.3",
    "diffusers==0.35.1",
    "tyro==0.9.35",
    "h5py",
    "fastapi",
    "uvicorn",
    "json_numpy",
    "datasets==3.6.0",
    "huggingface_hub==0.34.4",
    "jsonlines==4.0.0",
    "mujoco==3.3.5"
]


[project.optional-dependencies]
dev = [
    "black>=24.2.0",
    "gpustat",
    "ipython",
    "pre-commit",
    "ruff>=0.2.2",
]
sagemaker = [
    "boto3",
    "sagemaker"
]

[tool.setuptools.packages.find]
where = ["src"]
exclude = [
    "tests",         # test directory
    "cache",         # cache directory
    "playground",
    "results",       # result directory
    "checkpoints",   # model checkpoint
    "scripts",       # script directory
    "eval",          # evaluation related directory
    ".vscode",       # VS Code configuration directory
    "*.egg-info",    # egg-info metadata directory
    "**.log",         # log file
    "*.tmp",         # temporary file
    "*.bak",         # backup file
    "**__pycache__**",   # Python cache directory
    "*.pyc",         # compiled Python file
    "*.pyo",          # optimized Python file
    "**bar**",
    "unifolm_vla/config",

]

[build-system]
requires = ["setuptools>=65.6.3", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}


